Add files via upload
These is the files that I did in my project to the library system, as you can see I have a video presentation of my project, and I have a makefile that I used to build the system and compile an executable program. also catch2 for teste and the main code where is the library code system.